requirements.txt
Flask==2.3.2
Flask_SQLAlchemy==3.0.3
Werkzeug==2.3.7
pandas==2.2.3
python-dotenv==1.0.0
Jinja2==3.1.2
APScheduler==3.10.1
mysql-connector-python==8.1.0
SQLAlchemy==2.0.22
python-dateutil==2.8.2
xlsxwriter